Fixes the RoQ motion-comp 64-bit crash (main menu now works) and verifies the
core multiplayer paths: menu, networked connect to a dedicated server, 3D
render, and sound all run clean on the 64-bit build.

+++ src/client/cl_cin.c
@@ -93,7 +93,9 @@ typedef struct {
 	byte file[65536];
 	short sqrTable[256];
 
-	unsigned int mcomp[256];
+	intptr_t mcomp[256];   // 64-bit: motion-comp offsets are signed & added to a byte* (status[index] + mcomp).
+	                       // Was 'unsigned int': a negative offset zero-extended to a ~+4GB jump on x64 -> crash.
+	                       // (On 32-bit the unsigned wraparound happened to work.)
 	byte                *qStatus[2][32768];
 
 	long oldXOff, oldYOff, oldysize, oldxsize;
